// src/app/services/swagger/model/planetAbstractId.ts
/**
 * BoF REST API
 * Battle for honor interface
 *
 * OpenAPI spec version: 0.1.4
 * Contact: webmaster@batleforhonor.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * A container when only the database id if needed.
 */
export interface PlanetAbstractId { 
    /**
     * The database id.
     */
    idPlanet: number;
    /**
     * The name.
     */
    name: string;
    /**
     * .
     */
    isMain: boolean;
    /**
     * The database id.
     */
    idStarSystem: number;
}